Extract JWT handling from RestrictAccessAttribute
